+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Extender: Add Title field to change the Table of Contents title.
+
 ## [0.3.0] - 2022-12-19
 ### Changed
 - Change the module path to `go.abhg.dev/goldmark/toc`.