+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.0] - 2023-03-02
 ### Added
 - Extender: Add Title field to change the Table of Contents title.
 - Inspect: Add MaxDepth option to limit the depth of the Table of Contents.
+
+[0.4.0]: https://github.com/abhinav/goldmark-toc/releases/tag/v0.4.0
 
 ## [0.3.0] - 2022-12-19
 ### Changed